code-move
* move fields of `ActiveReplication` into `Instance`
* move methods of `ActiveReplication` into `Instance`
* move fields of `ActiveReplicationState` into `Instance`
* move methods of `ActiveReplicationState` into `Instance`
* move remaining code from replicated.rs into controller.rs
